requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
google-cloud-texttospeech
